refactor: let directory structure and routes follow the same pattern
