+++ src/views/instances/index.ts
@@ -1,3 +1,3 @@
-export { InstancesList } from './instances-list'
-export { NewInstance } from './new-instance'
-export { NewDhis2Instance } from './new-dhis2-instance'
+export { InstancesList } from './list/instances-list'
+export { NewInstance } from './new/new-instance'
+export { NewDhis2Instance } from './new-dhis2'
