+++ src/views/instances/list/reset-button.tsx
@@ -1,9 +1,9 @@
 import { useAlert } from '@dhis2/app-service-alerts'
 import { Button, IconClockHistory16 } from '@dhis2/ui'
 import React, { FC, useCallback, useState } from 'react'
-import { ConfirmationModal } from '../../components'
-import { useAuthAxios } from '../../hooks'
-import { Instance } from '../../types'
+import { ConfirmationModal } from '../../../components'
+import { useAuthAxios } from '../../../hooks'
+import { Instance } from '../../../types'
 
 type ResetButtonProps = {
     instanceId: number
