+++ src/views/instances/list/save-as-modal.module.css
@@ -2,7 +2,7 @@
     display: flex;
     flex-direction: row;
     gap: 12px;
-    align-items: center;
+    align-items: flex-end;
 }
 
 .container {
@@ -39,4 +39,10 @@
 
 .cancelButton:hover {
     color: var(--colors-blue700);
+}
+.field {
+    width: 480px;
+}
+.extension {
+    line-height: 40px;
 }
