+++ src/views/instances/new/new-instance.tsx
@@ -1,8 +1,8 @@
 import { Button, ButtonStrip, Card, Center, CircularLoader, NoticeBox, SingleSelectField, SingleSelectOption } from '@dhis2/ui'
 import { useCallback, useEffect, useRef, useState } from 'react'
 import { useNavigate } from 'react-router-dom'
-import { useAuthAxios } from '../../hooks'
-import { Instance, Stack } from '../../types'
+import { useAuthAxios } from '../../../hooks'
+import { Instance, Stack } from '../../../types'
 import styles from './new-instance.module.css'
 import { StackConfigurator } from './stack-configurator'
 import type { FC } from 'react'
