+++ src/views/instances/new/stack-configurator.module.css
