+++ src/views/instances/new/stack-configurator.tsx
@@ -2,8 +2,8 @@ import { Center, CheckboxField, CircularLoader, Divider, InputField, NoticeBox, 
 import cx from 'classnames'
 import type { Ref } from 'react'
 import { forwardRef, useEffect, useImperativeHandle, useState } from 'react'
-import { useAuthAxios } from '../../hooks'
-import { Group, Stack, StackParameter } from '../../types'
+import { useAuthAxios } from '../../../hooks'
+import { Group, Stack, StackParameter } from '../../../types'
 import { IMAGE_REPOSITORY, IMAGE_TAG, ParameterField } from './parameter-field'
 import styles from './stack-configurator.module.css'
 import fieldStyles from './parameter-field.module.css'
